fix: use uint32 keys for superchat tier colors and log unknown
Per audit reports/chat.md Q8. superchatTierColors used int64 keys for
values that are naturally 32-bit unsigned ARGB ints — small mismatch
but the wider type hid intent and would silently miscompare if any
value ever crossed the 2^31 boundary. Change map key type to uint32
and convert the parsed float64 via uint32.

Also log unknown bgColor values at debug level so tier-table drift
(YouTube adding or changing superchat colors) is visible without
polluting the error channel.

+++ internal/chat/types.go
@@ -66,8 +66,9 @@ type ChatResumeState struct {
 	RecentIDs         []string `json:"recentIds"`
 }
 
-// Superchat tier color mapping (YouTube's internal color codes).
-var superchatTierColors = map[int64]struct {
+// Superchat tier color mapping (YouTube's internal ARGB color codes).
+// Keys are uint32 since the raw values are 32-bit unsigned ARGB ints.
+var superchatTierColors = map[uint32]struct {
 	tier  int
 	color string
 }{
